WBYD-167 ディレクトリ位置調整
