--- contract/.gitignore
@@ -1,6 +1,0 @@
-node_modules
-
-#Hardhat files
-cache
-artifacts
-types
