--- contract/args/TestToken.js
@@ -1,1 +1,0 @@
-module.exports = ['100000000000000000000'];
