--- contract/package.json
@@ -1,29 +1,0 @@
-{
-  "name": "contract",
-  "scripts": {
-    "compile": "hardhat compile",
-    "test": "tsc --noEmit && hardhat test",
-    "update": "ncu -u"
-  },
-  "devDependencies": {
-    "@nomiclabs/hardhat-ethers": "^2.1.0",
-    "@nomiclabs/hardhat-etherscan": "^3.1.0",
-    "@nomiclabs/hardhat-waffle": "^2.0.3",
-    "@openzeppelin/contracts": "^4.7.0",
-    "@typechain/ethers-v5": "^10.1.0",
-    "@typechain/hardhat": "^6.1.2",
-    "@types/chai": "^4.3.0",
-    "@types/mocha": "^9.1.1",
-    "@types/node": "^17.0.42",
-    "chai": "^4.3.6",
-    "ethers": "^5.6.9",
-    "hardhat": "^2.10.0",
-    "hardhat-watcher": "^2.3.0",
-    "prettier": "^2.7.1",
-    "prettier-plugin-solidity": "^1.0.0-dev.22",
-    "solhint": "^3.3.7",
-    "ts-node": "^10.8.2",
-    "typechain": "^8.0.0",
-    "typescript": "^4.7.4"
-  }
-}
