--- contract/tsconfig.json
@@ -1,14 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "es2018",
-    "module": "commonjs",
-    "strict": true,
-    "esModuleInterop": true,
-    "outDir": "dist",
-    "declaration": true,
-    "types": ["node", "mocha"]
-  },
-  "include": ["./scripts", "./types", "./tests"],
-  "exclude": ["node_modules"],
-  "files": ["./hardhat.config.ts"]
-}
