+++ package.json
@@ -5,7 +5,7 @@
     "./packages/*"
   ],
   "scripts": {
-    "compile": "lerna run compile --scope=contract",
+    "compile": "lerna run compile --scope=@wbydo-blockchain-project/contracts",
     "test": "lerna run test --scope=@wbydo-blockchain-project/contracts"
   },
   "devDependencies": {
