+++ packages/contracts/.gitignore
@@ -2,8 +2,8 @@ node_modules
 .env
 coverage
 coverage.json
-typechain
-typechain-types
+dist
+
 
 #Hardhat files
 cache
