+++ packages/contracts/.solhint.json
@@ -1,7 +1,7 @@
 {
   "extends": "solhint:recommended",
   "rules": {
-    "compiler-version": ["error", "^0.8.0"],
+    "compiler-version": ["error", "^0.8.13"],
     "func-visibility": ["warn", { "ignoreConstructors": true }]
   }
 }
