--- packages/contracts/contracts/contracts/TestToken.sol
@@ -1,10 +1,0 @@
-//SPDX-License-Identifier: Unlicense
-pragma solidity 0.8.9;
-
-import "@openzeppelin/contracts/token/ERC20/ERC20.sol";
-
-contract TestToken is ERC20 {
-    constructor(uint256 amount) ERC20("", "TEST") {
-        _mint(_msgSender(), amount);
-    }
-}
