+++ packages/contracts/scripts/README.md
@@ -1,0 +1,2 @@
+TODO:
+管理画面を作ったら消去予定
