+++ packages/contracts/scripts/mint.ts
