+++ packages/contracts/tests/WbydoProfileNft.test.ts
@@ -1,9 +1,9 @@
+import { loadFixture } from '@nomicfoundation/hardhat-network-helpers';
+
 import { expect } from 'chai';
-import { waffle, ethers } from 'hardhat';
+import { ethers } from 'hardhat';
 
 import { BigNumber } from 'ethers';
-
-const { loadFixture } = waffle;
 
 const baseURI = 'https://example.com/metadata/';
