+++ packages/contracts/tsconfig.json
@@ -6,5 +6,8 @@
     "forceConsistentCasingInFileNames": true,
     "strict": true,
     "skipLibCheck": true
-  }
+  },
+  "include": ["./scripts", "./types", "./tests"],
+  "exclude": ["node_modules"],
+  "files": ["./hardhat.config.ts"]
 }
