--- workspace.code-workspace
@@ -1,14 +1,0 @@
-{
-  "folders": [
-    {
-      "path": "frontend"
-    },
-    {
-      "path": "contract"
-    },
-    {
-      "path": "."
-    }
-  ],
-  "settings": {}
-}
